refactor: just moved the components representing elements in the elements folder
