+++ client/src/app/components/elements/item-card-borrow/item-card-borrow.component.scss
