+++ client/src/app/components/elements/item-card-lend/item-card-lend.component.ts
