+++ client/src/app/components/elements/item-card-search/item-card-search.component.spec.ts
