+++ client/src/app/components/elements/navbar/navbar.component.scss
@@ -1,4 +1,4 @@
-@import "../../../variables.scss";
+@import "../../../../variables.scss";
 
 .navbar {
   height: $navbar-height;
